feat: create admin in seeder
