+++ example.env
@@ -4,3 +4,6 @@ EMAIL_PASS=password
 JWT_SECRET=exmapleForJWTSecret
 PORT=3000
 FORGET_PASSWORD_SECRET=example
+SU_USERNAME=admin_username
+SU_EMAIL=admin_email
+SU_PASS=admin_password
